refactor: update blog api delete

+++ part4/bloglist/controllers/blogs.js
@@ -17,8 +17,8 @@ blogsRouter.post("/", async (req, res) => {
   if (!decodedToken.id) {
     return res.status(401).json({ error: "token invalid" });
   }
-  const user = await User.findById(decodedToken.id);
 
+  const user = await User.findById(decodedToken.id);
   if (!user) {
     return res.status(400).json({
       error: "userId missing or not valid",
@@ -46,8 +46,24 @@ blogsRouter.post("/", async (req, res) => {
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
-  const deletedBlog = await Blog.findByIdAndDelete(req.params.id);
-  res.status(204).json(deletedBlog);
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" });
+  }
+
+  const blog = await Blog.findById(req.params.id);
+  if (!blog) {
+    return res.status(404).json({
+      error: "blog not found",
+    });
+  }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return res.status(403).json({ error: "only creater can delete this blog" });
+  }
+
+  await Blog.findByIdAndDelete(req.params.id);
+  res.status(204).end();
 });
 
 blogsRouter.put("/:id", async (req, res) => {
